python/requirements.txt
# PDF Processing - STABLE VERSIONS
pdfplumber==0.10.3
PyPDF2==3.0.1

# Vector Database & Embeddings - STABLE
faiss-cpu==1.7.4
sentence-transformers==2.2.2

# Core ML/NLP - STABLE
numpy==1.24.3
scikit-learn==1.3.2
torch==2.0.1
transformers==4.30.2

# API & Web Framework - STABLE
flask==2.3.3
flask-cors==4.0.0
pydantic==2.3.0

# Security & Authentication - STABLE
pyjwt==2.8.0
bcrypt==4.0.1
python-dotenv==1.0.0

# Utilities - STABLE
python-dateutil==2.8.2
requests==2.31.0

# Testing - STABLE
pytest==7.4.3
pytest-cov==4.1.0

# Additional stable packages
werkzeug==2.3.7
click==8.1.7
